requirements.txt
# Core Flask dependencies
Flask>=3.0.0
Werkzeug>=3.0.0
Jinja2>=3.1.0

# Internationalization
Flask-Babel==4.0.0
Babel==2.14.0

# Desktop GUI
pywebview>=4.0.0

# Cryptography
cryptography>=41.0.0

# SSH/SFTP
paramiko>=3.0.0

# HTTP requests
requests>=2.31.0
urllib3>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-flask>=1.2.0
pytest-cov>=4.1.0

# Code quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# Security
bandit>=1.7.0

# Build tools
pyinstaller>=5.13.0 